marccd: Throw exception on startCount if camera is busy
Previously, startCount waited for a short time before starting
acquisition for the camera to report idle state. After a timeout,
the method would ignore the busy state and start acquisition anyway.
This patch changes this behavior to raise a runtime exception if
the camera is busy. This is more reliable generally. Now, users
must call waitForIdle if they require sequential acquisitions.

Note that this patch must be checked against bugs in MarCCD server
software, so it may be reverted or amended if necessary.

git-svn-id: https://svn.lnls.br/LNLS/Py4Syn/trunk@242 8c29a70a-4b4f-4fad-ae86-83d8b3214d8e