marccd: Wait until the camera reports acquisition started
To make measurement more precise, wait until the camera reports
the acquiring state. In general, the camera will report the
acquiring state after around 50ms, which is reasonable. Some
times, however, it may take up to 300ms to report it. The method
waitForIdle may help to move this delay to a safe moment.

git-svn-id: https://svn.lnls.br/LNLS/Py4Syn/trunk@244 8c29a70a-4b4f-4fad-ae86-83d8b3214d8e