docs: Remove malfunctioning warning from Kepco BOP documentation
The reason for malfunctioning was found out to be caused by a configuration.
Kepco requires two jumpers to be correctly configured for software based
operation. Previously it was configured with only one jumper, which
sometimes resulted in noise present in analog inputs to be interpreted
as analog commands.

From: Henrique Dante de Almeida <henrique.almeida@lnls.br>

git-svn-id: https://svn.lnls.br/LNLS/Py4Syn/trunk@281 8c29a70a-4b4f-4fad-ae86-83d8b3214d8e